[sql] changed type of money to decimal (15,2)

+++ sql/create_table.sql
@@ -123,7 +123,7 @@ CREATE TABLE IF NOT EXISTS `meatballs`.`menu_item`
 (
   `mitem_id` INTEGER PRIMARY KEY AUTO_INCREMENT,
   `category` CHAR(45) NULL,
-  `price` DOUBLE NULL,
+  `price` DECIMAL(15,2) NULL,
   `name` VARCHAR(65) NULL,
   `image` VARCHAR(95) NULL
 )
